reorganized folder structure and normalized namespace conventions
