+++ CSSBot/Services/Counters/Commands/CounterCommands.cs
