+++ CSSBot/Services/Counters/CounterService.cs
