+++ CSSBot/Services/Counters/Models/Counter.cs
