+++ CSSBot/Services/Counters/Models/CounterFile.cs
