+++ CSSBot/Services/Reminders/Commands/ReminderCommands.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
+using CSSBot.Reminders.Models;
 
 namespace CSSBot
 {
